Moved car code to vehicle class and made car subclass of vehicle
